Changed file and code structure
